[Project]: changed project file strucutre - moved core code to 'core/' dir, moved 'theming' code to 'theming/'.
