+++ src/ts/core/helpers.ts
