+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
     "files": [
-        "src/ts/lib/json2html.ts",
+        "src/ts/core/json2html.ts",
         "src/ts/app.ts",
     ],
     "compilerOptions": {
