+++ webpack.config.js
@@ -5,7 +5,7 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 
 module.exports = {
     entry: {
-      'lib/json2html': './src/ts/lib/json2html.ts',
+      'lib/json2html': './src/ts/core/json2html.ts',
       'app': './src/ts/app.ts',
     },
